# Build abstract and portable SIMD using xsimd