--- a/velox/dwio/dwrf/common/DecoderUtil.cpp
+++ b/velox/dwio/dwrf/common/DecoderUtil.cpp
@@ -20,8 +20,6 @@
 
 namespace facebook::velox::dwrf {
 
-using V32 = simd::Vectors<int32_t>;
-
 int32_t nonNullRowsFromDense(
     const uint64_t* nulls,
     int32_t numRows,
@@ -101,15 +99,15 @@ bool nonNullRowsFromSparse(
         continue;
       }
       auto numNonNull = __builtin_popcount(flags);
-      auto setBits = V32::compareSetBits(flags);
       // contiguous inner rows.
       auto innerStart = innerFor(i);
-      V32::store(inner + numInner, V32::iota() + innerStart);
+      (simd::iota<int32_t>() + innerStart).store_unaligned(inner + numInner);
       if (isFilter) {
-        simd::storePermute(
-            outer + numInner, V32::load(rows.data() + i), setBits);
+        simd::filter(xsimd::load_unaligned(rows.data() + i), flags)
+            .store_unaligned(outer + numInner);
       } else {
-        V32::store(outer + numInner, setBits + i);
+        (detail::bitMaskIndices<int32_t>(flags) + i)
+            .store_unaligned(outer + numInner);
       }
       // We processed 'width' consecutive. The inner count is incremented
       // by number of non-nulls. Nulls are counted for the 'width' rows,
@@ -121,7 +119,7 @@ bool nonNullRowsFromSparse(
       lastNonNull = rows[i + width - 1];
       numNulls += width - numNonNull;
     } else {
-      auto next8Rows = V32::load(rows.data() + i);
+      auto next8Rows = xsimd::load_unaligned(rows.data() + i);
       uint16_t flags = simd::gather8Bits(nulls, next8Rows, width);
       if (outputNulls) {
         resultNullBytes[i / 8] = flags;
@@ -130,14 +128,14 @@ bool nonNullRowsFromSparse(
       if (!flags) {
         continue;
       }
-      auto setBits = V32::compareSetBits(flags);
       if (isFilter) {
         // The non-null indices among 'rows' are possible filter results.
-        simd::storePermute(outer + numInner, next8Rows, setBits);
+        simd::filter(next8Rows, flags).store_unaligned(outer + numInner);
       } else {
         // Make scatter indices so that there are gaps for 'rows' that hit a
         // null.
-        V32::store(outer + numInner, setBits + i);
+        (detail::bitMaskIndices<int32_t>(flags) + i)
+            .store_unaligned(outer + numInner);
       }
       // Calculate the inner row corresponding to each non-null in 'next8Rows'.
       while (flags) {