… CMakeLists.txt
… velox/buffer/CMakeLists.txt
… velox/common/base/CMakeLists.txt
… velox/common/base/SimdUtil-inl.h
… velox/common/base/SimdUtil.h
… velox/common/base/tests/RawVectorTest.cpp
… velox/common/base/tests/SimdUtilTest.cpp
… velox/dwio/common/CMakeLists.txt
… velox/dwio/common/tests/CMakeLists.txt
… velox/dwio/dwrf/reader/ColumnVisitors.h
… velox/dwio/dwrf/test/DecoderUtilTest.cpp
… velox/exec/HashTable.h
… velox/exec/VectorHasher-inl.h
… velox/exec/VectorHasher.cpp
… velox/exec/VectorHasher.h
… velox/experimental/codegen/code_generator/CMakeLists.txt
… velox/experimental/codegen/transform/utils/CMakeLists.txt
… velox/experimental/codegen/utils/timer/tests/CMakeLists.txt
… velox/experimental/codegen/utils/timer/tests/ClockTests.cpp
… velox/experimental/codegen/utils/timer/tests/TARGETS_disabled
… velox/functions/prestosql/types/CMakeLists.txt
… velox/row/CMakeLists.txt
… velox/type/Filter.cpp
… velox/type/Filter.h
… velox/type/tests/FilterBenchmark.cpp
… velox/type/tests/FilterTest.cpp
… velox/vector/BiasVector-inl.h
… velox/vector/BiasVector.h
… velox/vector/CMakeLists.txt
… velox/vector/ConstantVector.h