Summary: This change replaces all Intel intrinsics with portable xsimd wrappers.  It only runs on AVX2 chips for now.  Subsequent changes will make it portable to SSE4.2 and NEON.

Differential Revision: D35579863