@Yuhta 2022-04-21
Below are some of the more representative benchmarks.  There are also some benchmarks with vectors but the operations are simple and the results are long so I did not include it here.  I did not find any benchmark for DWRF, but the operations are similar to `Filter`.  All the dispatches are static and can be inlined so I am not worried too much about the performance on AVX2.

Baseline:

```
$ buck run @mode/opt //velox/type/tests:velox_filter_benchmark
============================================================================
velox/type/tests/FilterBenchmark.cpp     relative  time/iter   iters/s
============================================================================
scalarDense                                                10.94ms     91.44
simdDense                                        166.8%     6.56ms    152.52
scalarSparse                                                2.96ms    337.70
simdSparse                                      204.12%     1.45ms    689.30

$ buck run @mode/opt //velox/exec/benchmarks:vector_hasher_benchmark
============================================================================
[...]/benchmarks/VectorHasherBenchmark.cpp     relative  time/iter   iters/s
============================================================================
computeValueIdsBigintNoNulls                                3.70ms    270.08
computeValueIdsBigintWithNulls                  12.652%    29.26ms     34.17
computeValueIdsIntegerNoNulls                               3.46ms    288.87
computeValueIdsIntegerWithNulls                 10.902%    31.75ms     31.49
computeValueIdsSmallintNoNulls                              2.59ms    386.02
computeValueIdsSmallintWithNulls                 8.179%    31.67ms     31.57
computeValueIdsDictionaryStrings                           88.11ms     11.35
computeValueIdsFlatStrings                       5.938%      1.48s   673.90m
```

New version:
```
$ buck run @mode/opt //velox/type/tests:velox_filter_benchmark
============================================================================
velox/type/tests/FilterBenchmark.cpp     relative  time/iter   iters/s
============================================================================
scalarDense                                                11.32ms     88.33
simdDense                                       169.76%     6.67ms    149.95
scalarSparse                                                3.13ms    319.22
simdSparse                                      222.47%     1.41ms    710.15

$ buck run @mode/opt //velox/exec/benchmarks:vector_hasher_benchmark
============================================================================
[...]/benchmarks/VectorHasherBenchmark.cpp     relative  time/iter   iters/s
============================================================================
computeValueIdsBigintNoNulls                                4.32ms    231.54
computeValueIdsBigintWithNulls                  13.523%    31.94ms     31.31
computeValueIdsIntegerNoNulls                               3.98ms    251.56
computeValueIdsIntegerWithNulls                  11.74%    33.86ms     29.53
computeValueIdsSmallintNoNulls                              2.86ms    349.33
computeValueIdsSmallintWithNulls                9.0541%    31.62ms     31.63
computeValueIdsDictionaryStrings                           91.70ms     10.91
computeValueIdsFlatStrings                       5.844%      1.57s   637.33m
```