@pedroerp 2022-04-21
> I did not find any benchmark for DWRF (...)

I'm planning to add some micro-benchmark for dwrf encoding/decoding as part of the micro-benchmark CI tests. Will probably get to that sometime next week or so.