@tanjialiang 2022-05-02
Hi Jimmy, per Orri's request I helped run some tests for this PR on my Intel chip Macbook. Here is what I got for velox_cache_test


SSD: Ssd cache IO: Write 1109MB read 631MB Size 0GB Occupied 0GB7K entries.
GroupStats: <dummy FileGroupStats>
[       OK ] AsyncDataCacheTest.ssd (21423 ms)
[----------] 4 tests from AsyncDataCacheTest (25355 ms total)

[----------] 1 test from SsdFileTest
[ RUN      ] SsdFileTest.writeAndRead
/Users/jtan6/git-repos/presto_cpp/velox/velox/common/caching/tests/SsdFileTest.cpp:201: Failure
Value of: ssdFile_->find(RawFileCacheKey{fileName_.id(), ssdSize}).empty()
  Actual: true
Expected: false
/Users/jtan6/git-repos/presto_cpp/velox/velox/common/caching/tests/SsdFileTest.cpp:168: Failure
Value of: ssdPins.back().empty()
  Actual: true
Expected: false
/Users/jtan6/git-repos/presto_cpp/velox/velox/common/caching/tests/SsdFileTest.cpp:168: Failure
Value of: ssdPins.back().empty()
  Actual: true
Expected: false
/Users/jtan6/git-repos/presto_cpp/velox/velox/common/caching/tests/SsdFileTest.cpp:168: Failure
Value of: ssdPins.back().empty()
  Actual: true
Expected: false
/Users/jtan6/git-repos/presto_cpp/velox/velox/common/caching/tests/SsdFileTest.cpp:168: Failure
Value of: ssdPins.back().empty()
  Actual: true
Expected: false
/Users/jtan6/git-repos/presto_cpp/velox/velox/common/caching/tests/SsdFileTest.cpp:168: Failure
Value of: ssdPins.back().empty()
  Actual: true
Expected: false
E20220502 11:43:35.497864 260736512 Exceptions.h:68] Line: /Users/jtan6/git-repos/presto_cpp/velox/velox/common/caching/SsdFile.cpp:182, Function:load, Expression: runSize >= entry->size()(1 vs. 4096) IOERR SSd cache entry shorter than requested range, Source: RUNTIME, ErrorCode: INVALID_STATE
unknown file: Failure
C++ exception with description "Exception: VeloxRuntimeError
Error Source: RUNTIME
Error Code: INVALID_STATE
Reason: (1 vs. 4096) IOERR SSd cache entry shorter than requested range
Retriable: False
Expression: runSize >= entry->size()
Function: load
File: /Users/jtan6/git-repos/presto_cpp/velox/velox/common/caching/SsdFile.cpp
Line: 182
Stack trace:
Stack trace has been disabled. Use --velox_exception_stacktrace=true to enable it.
" thrown in the test body.
[  FAILED  ] SsdFileTest.writeAndRead (10589 ms)
[----------] 1 test from SsdFileTest (10589 ms total)

[----------] 1 test from SsdFileTrackerTest
[ RUN      ] SsdFileTrackerTest.tracker
[       OK ] SsdFileTrackerTest.tracker (1 ms)
[----------] 1 test from SsdFileTrackerTest (1 ms total)

[----------] Global test environment tear-down
[==========] 8 tests from 4 test suites ran. (36013 ms total)
[  PASSED  ] 7 tests.
[  FAILED  ] 1 test, listed below:
[  FAILED  ] SsdFileTest.writeAndRead

 1 FAILED TEST