@Yuhta 2022-05-02
Hi @tanjialiang , can you try it on base version?  The code in `SsdFileTest` does not seem related to SIMD.