@oerling 2022-05-02
Jialiang just tried it. It does not work on main either.

 

 

From: Jimmy Lu ***@***.***> 
Sent: Monday, May 2, 2022 12:17 PM
To: facebookincubator/velox ***@***.***>
Cc: oerling ***@***.***>; Review requested ***@***.***>
Subject: Re: [facebookincubator/velox] Build abstract and portable SIMD using xsimd (PR #1445)

 

Hi @tanjialiang <https://github.com/tanjialiang>  , can you try it on base version? The code in SsdFileTest does not seem related to SIMD.

—
Reply to this email directly, view it on GitHub <https://github.com/facebookincubator/velox/pull/1445#issuecomment-1115266821> , or unsubscribe <https://github.com/notifications/unsubscribe-auth/AKPPPT5S2HIJWKIBNXX3F33VIASZ5ANCNFSM5TZB34XQ> .
You are receiving this because your review was requested.Message ID: ***@***.***>